- Added placeholder for game environment page
